Support for virtual (predefined) screen space. ref #664
* The defined size in game.project is used as a virtual screen space (origo in bottom-left)
* The actual window size is used to render gui/debug texts
* Data is propagated from engine to graphics and gui contexts
* Window resize updates gui context
* Input is transformed to virtual screen space in engine, then propagated
* Positional input is represented as floats due to scaling
* Physical screen space is read back for xwindows and win32
* Fixed compile-warnings
* In gui, ref-scale is applied to scaling, taking the minimum of the x-y factors for uniform scaling
* Fixed guidemo
* Updated tests
* Exposed predefined size in graphics and render scripts

+++ gameobject/src/gameobject/gameobject.h
@@ -100,14 +100,14 @@ namespace dmGameObject
         dmhash_t m_ActionId;
         /// Value of the input [0,1]
         float m_Value;
-        /// Cursor X coordinate
-        int32_t m_X;
-        /// Cursor Y coordinate
-        int32_t m_Y;
-        /// Cursor dx since last frame
-        int32_t m_DX;
-        /// Cursor dy since last frame
-        int32_t m_DY;
+        /// Cursor X coordinate, in virtual screen space
+        float m_X;
+        /// Cursor Y coordinate, in virtual screen space
+        float m_Y;
+        /// Cursor dx since last frame, in virtual screen space
+        float m_DX;
+        /// Cursor dy since last frame, in virtual screen space
+        float m_DY;
         /// If the input was 0 last update
         uint16_t m_Pressed : 1;
         /// If the input turned from above 0 to 0 this update
